feat: --sources and --files can now refer to directories
